extract shared config and use environment variables
